move gpu-allocator stuff into it's own file
