+++ wgpu-hal/src/dx12/device.rs
@@ -3,30 +3,25 @@ use crate::{
     FormatAspects,
 };
 
-use super::{conv, descriptor, view};
+use super::{
+    conv, descriptor, view,
+    windows_rs_suballocation::{
+        create_allocator_wrapper, free_buffer_allocation, free_texture_allocation,
+    },
+};
 use parking_lot::Mutex;
 use std::{ffi, mem, num::NonZeroU32, ptr, slice, sync::Arc};
 use winapi::{
     shared::{dxgiformat, dxgitype, minwindef::BOOL, winerror},
-    um::{d3d12, d3dcompiler, synchapi, winbase},
+    um::{
+        d3d12::{self, D3D12_RESOURCE_DESC},
+        d3dcompiler, synchapi, winbase,
+    },
     Interface,
-};
-
-#[cfg(not(feature = "windows_rs"))]
-use winapi::um::d3d12::D3D12_RESOURCE_DESC;
-
-#[cfg(feature = "windows_rs")]
-use gpu_allocator::{
-    d3d12::{winapi_d3d12::D3D12_RESOURCE_DESC, AllocationCreateDesc, ToWinapi, ToWindows},
-    MemoryLocation,
 };
 
 // this has to match Naga's HLSL backend, and also needs to be null-terminated
 const NAGA_LOCATION_SEMANTIC: &[u8] = b"LOC\0";
-// TODO: find the exact value
-// TODO: figure out if this is even needed?
-#[allow(unused)] // TODO: Exists until windows-rs is standard, then it can probably be removed?
-const D3D12_HEAP_FLAG_CREATE_NOT_ZEROED: u32 = d3d12::D3D12_HEAP_FLAG_NONE;
 
 impl super::Device {
     pub(super) fn new(
@@ -35,18 +30,7 @@ impl super::Device {
         private_caps: super::PrivateCapabilities,
         library: &Arc<native::D3D12Lib>,
     ) -> Result<Self, crate::DeviceError> {
-        #[cfg(feature = "windows_rs")]
-        let mem_allocator = {
-            let device = raw.as_ptr();
-
-            match gpu_allocator::d3d12::Allocator::new(&gpu_allocator::d3d12::AllocatorCreateDesc {
-                device: device.as_windows().clone(),
-                debug_settings: Default::default(),
-            }) {
-                Ok(allocator) => allocator,
-                Err(e) => panic!("Failed to create d3d12 allocator, error: {}", e),
-            }
-        };
+        let mem_allocator = create_allocator_wrapper(&raw)?;
 
         let mut idle_fence = native::Fence::null();
         let hr = unsafe {
@@ -189,8 +173,7 @@ impl super::Device {
             #[cfg(feature = "renderdoc")]
             render_doc: Default::default(),
             null_rtv_handle,
-            #[cfg(feature = "windows_rs")]
-            mem_allocator: Mutex::new(mem_allocator),
+            mem_allocator,
         })
     }
 
@@ -338,14 +321,13 @@ impl super::Device {
             size,
             mip_level_count,
             sample_count,
-            #[cfg(feature = "windows_rs")]
             allocation: None,
         }
     }
 }
 
 impl crate::Device<super::Api> for super::Device {
-    unsafe fn exit(self, queue: super::Queue) {
+    unsafe fn exit(mut self, queue: super::Queue) {
         self.rtv_pool.lock().free_handle(self.null_rtv_handle);
         unsafe { self.rtv_pool.into_inner().destroy() };
         unsafe { self.dsv_pool.into_inner().destroy() };
@@ -353,14 +335,16 @@ impl crate::Device<super::Api> for super::Device {
         unsafe { self.sampler_pool.into_inner().destroy() };
         unsafe { self.shared.destroy() };
         unsafe { self.idler.destroy() };
+        drop(self.mem_allocator.take());
         unsafe { queue.raw.destroy() };
     }
 
-    #[cfg(feature = "windows_rs")]
     unsafe fn create_buffer(
         &self,
         desc: &crate::BufferDescriptor,
     ) -> Result<super::Buffer, crate::DeviceError> {
+        use super::windows_rs_suballocation::my_buffer_hr;
+
         let mut resource = native::Resource::null();
         let mut size = desc.size;
         if desc.usage.contains(crate::BufferUses::UNIFORM) {
@@ -370,7 +354,6 @@ impl crate::Device<super::Api> for super::Device {
 
         let raw_desc = D3D12_RESOURCE_DESC {
             Dimension: d3d12::D3D12_RESOURCE_DIMENSION_BUFFER,
-            // TODO: Alignment, see https://www.asawicki.info/news_1726_secrets_of_direct3d_12_resource_alignment
             Alignment: 0,
             Width: size,
             Height: 1,
@@ -385,36 +368,7 @@ impl crate::Device<super::Api> for super::Device {
             Flags: conv::map_buffer_usage_to_resource_flags(desc.usage),
         };
 
-        let is_cpu_read = desc.usage.contains(crate::BufferUses::MAP_READ);
-        let is_cpu_write = desc.usage.contains(crate::BufferUses::MAP_WRITE);
-        // TODO: These are probably wrong?
-        let location = match (is_cpu_read, is_cpu_write) {
-            (true, true) => MemoryLocation::CpuToGpu,
-            (true, false) => MemoryLocation::GpuToCpu,
-            (false, true) => MemoryLocation::CpuToGpu,
-            (false, false) => MemoryLocation::GpuOnly,
-        };
-
-        let mut allocator = self.mem_allocator.lock();
-        let allocation_desc = AllocationCreateDesc::from_winapi_d3d12_resource_desc(
-            allocator.device().as_winapi(),
-            &raw_desc,
-            "Example allocation",
-            location,
-        );
-        let allocation = allocator.allocate(&allocation_desc)?;
-
-        let hr = unsafe {
-            self.raw.CreatePlacedResource(
-                allocation.heap().as_winapi() as *mut _,
-                allocation.offset(),
-                &raw_desc,
-                d3d12::D3D12_RESOURCE_STATE_COMMON,
-                ptr::null(),
-                &d3d12::ID3D12Resource::uuidof(),
-                resource.mut_void(),
-            )
-        };
+        let (hr, allocation) = my_buffer_hr(self, desc, raw_desc, &mut resource)?;
 
         hr.into_device_result("Buffer creation")?;
         if let Some(label) = desc.label {
@@ -425,104 +379,18 @@ impl crate::Device<super::Api> for super::Device {
         Ok(super::Buffer {
             resource,
             size,
-            allocation: Some(allocation),
+            allocation,
         })
     }
 
-    #[cfg(not(feature = "windows_rs"))]
-    unsafe fn create_buffer(
-        &self,
-        desc: &crate::BufferDescriptor,
-    ) -> Result<super::Buffer, crate::DeviceError> {
-        let mut resource = native::Resource::null();
-        let mut size = desc.size;
-        if desc.usage.contains(crate::BufferUses::UNIFORM) {
-            let align_mask = d3d12::D3D12_CONSTANT_BUFFER_DATA_PLACEMENT_ALIGNMENT as u64 - 1;
-            size = ((size - 1) | align_mask) + 1;
-        }
-
-        let raw_desc = D3D12_RESOURCE_DESC {
-            Dimension: d3d12::D3D12_RESOURCE_DIMENSION_BUFFER,
-            Alignment: 0,
-            Width: size,
-            Height: 1,
-            DepthOrArraySize: 1,
-            MipLevels: 1,
-            Format: dxgiformat::DXGI_FORMAT_UNKNOWN,
-            SampleDesc: dxgitype::DXGI_SAMPLE_DESC {
-                Count: 1,
-                Quality: 0,
-            },
-            Layout: d3d12::D3D12_TEXTURE_LAYOUT_ROW_MAJOR,
-            Flags: conv::map_buffer_usage_to_resource_flags(desc.usage),
-        };
-
-        let is_cpu_read = desc.usage.contains(crate::BufferUses::MAP_READ);
-        let is_cpu_write = desc.usage.contains(crate::BufferUses::MAP_WRITE);
-
-        let heap_properties = d3d12::D3D12_HEAP_PROPERTIES {
-            Type: d3d12::D3D12_HEAP_TYPE_CUSTOM,
-            CPUPageProperty: if is_cpu_read {
-                d3d12::D3D12_CPU_PAGE_PROPERTY_WRITE_BACK
-            } else if is_cpu_write {
-                d3d12::D3D12_CPU_PAGE_PROPERTY_WRITE_COMBINE
-            } else {
-                d3d12::D3D12_CPU_PAGE_PROPERTY_NOT_AVAILABLE
-            },
-            MemoryPoolPreference: match self.private_caps.memory_architecture {
-                super::MemoryArchitecture::NonUnified if !is_cpu_read && !is_cpu_write => {
-                    d3d12::D3D12_MEMORY_POOL_L1
-                }
-                _ => d3d12::D3D12_MEMORY_POOL_L0,
-            },
-            CreationNodeMask: 0,
-            VisibleNodeMask: 0,
-        };
-
-        let hr = unsafe {
-            self.raw.CreateCommittedResource(
-                &heap_properties,
-                if self.private_caps.heap_create_not_zeroed {
-                    D3D12_HEAP_FLAG_CREATE_NOT_ZEROED
-                } else {
-                    d3d12::D3D12_HEAP_FLAG_NONE
-                },
-                &raw_desc,
-                d3d12::D3D12_RESOURCE_STATE_COMMON,
-                ptr::null(),
-                &d3d12::ID3D12Resource::uuidof(),
-                resource.mut_void(),
-            )
-        };
-
-        hr.into_device_result("Buffer creation")?;
-        if let Some(label) = desc.label {
-            let cwstr = conv::map_label(label);
-            unsafe { resource.SetName(cwstr.as_ptr()) };
-        }
-
-        Ok(super::Buffer { resource, size })
-    }
-
-    #[cfg(feature = "windows_rs")]
     unsafe fn destroy_buffer(&self, mut buffer: super::Buffer) {
-        // TODO: Destroy or Release here?
         unsafe { buffer.resource.destroy() };
+        // Only happens when it's using the windows_rs feature and there's an allocation
         if let Some(alloc) = buffer.allocation.take() {
-            match self.mem_allocator.lock().free(alloc) {
-                Ok(_) => (),
-                // TODO: Don't panic here
-                Err(e) => panic!("failed to destroy dx12 buffer, {}", e),
-            };
+            free_buffer_allocation(alloc, self.mem_allocator.as_ref().unwrap());
         }
     }
 
-    #[cfg(not(feature = "windows_rs"))]
-    unsafe fn destroy_buffer(&self, buffer: super::Buffer) {
-        unsafe { buffer.resource.destroy() };
-    }
-
-    #[cfg(feature = "windows_rs")]
     unsafe fn map_buffer(
         &self,
         buffer: &super::Buffer,
@@ -541,31 +409,6 @@ impl crate::Device<super::Api> for super::Device {
         })
     }
 
-    #[cfg(not(feature = "windows_rs"))]
-    unsafe fn map_buffer(
-        &self,
-        buffer: &super::Buffer,
-        range: crate::MemoryRange,
-    ) -> Result<crate::BufferMapping, crate::DeviceError> {
-        let mut ptr = ptr::null_mut();
-        let hr = unsafe { (*buffer.resource).Map(0, ptr::null(), &mut ptr) };
-        hr.into_device_result("Map buffer")?;
-        Ok(crate::BufferMapping {
-            ptr: ptr::NonNull::new(unsafe { ptr.offset(range.start as isize).cast::<u8>() })
-                .unwrap(),
-            //TODO: double-check this. Documentation is a bit misleading -
-            // it implies that Map/Unmap is needed to invalidate/flush memory.
-            is_coherent: true,
-        })
-    }
-
-    #[cfg(feature = "windows_rs")]
-    unsafe fn unmap_buffer(&self, buffer: &super::Buffer) -> Result<(), crate::DeviceError> {
-        unsafe { (*buffer.resource).Unmap(0, ptr::null()) };
-        Ok(())
-    }
-
-    #[cfg(not(feature = "windows_rs"))]
     unsafe fn unmap_buffer(&self, buffer: &super::Buffer) -> Result<(), crate::DeviceError> {
         unsafe { (*buffer.resource).Unmap(0, ptr::null()) };
         Ok(())
@@ -574,11 +417,12 @@ impl crate::Device<super::Api> for super::Device {
     unsafe fn flush_mapped_ranges<I>(&self, _buffer: &super::Buffer, _ranges: I) {}
     unsafe fn invalidate_mapped_ranges<I>(&self, _buffer: &super::Buffer, _ranges: I) {}
 
-    #[cfg(feature = "windows_rs")]
     unsafe fn create_texture(
         &self,
         desc: &crate::TextureDescriptor,
     ) -> Result<super::Texture, crate::DeviceError> {
+        use super::windows_rs_suballocation::my_texture_hr;
+
         let mut resource = native::Resource::null();
 
         let raw_desc = D3D12_RESOURCE_DESC {
@@ -610,28 +454,7 @@ impl crate::Device<super::Api> for super::Device {
             Flags: conv::map_texture_usage_to_resource_flags(desc.usage),
         };
 
-        // TODO: What memory location should the texture be stored in?
-        let location = MemoryLocation::Unknown;
-        let mut allocator = self.mem_allocator.lock();
-        let allocation_desc = AllocationCreateDesc::from_winapi_d3d12_resource_desc(
-            allocator.device().as_winapi(),
-            &raw_desc,
-            "Example allocation",
-            location,
-        );
-        let allocation = allocator.allocate(&allocation_desc)?;
-
-        let hr = unsafe {
-            self.raw.CreatePlacedResource(
-                allocation.heap().as_winapi() as *mut _,
-                allocation.offset(),
-                &raw_desc,
-                d3d12::D3D12_RESOURCE_STATE_COMMON,
-                ptr::null(), // clear value
-                &d3d12::ID3D12Resource::uuidof(),
-                resource.mut_void(),
-            )
-        };
+        let (hr, allocation) = my_texture_hr(self, raw_desc, &mut resource)?;
 
         hr.into_device_result("Texture creation")?;
         if let Some(label) = desc.label {
@@ -646,103 +469,15 @@ impl crate::Device<super::Api> for super::Device {
             size: desc.size,
             mip_level_count: desc.mip_level_count,
             sample_count: desc.sample_count,
-            allocation: Some(allocation),
+            allocation,
         })
     }
 
-    #[cfg(not(feature = "windows_rs"))]
-    unsafe fn create_texture(
-        &self,
-        desc: &crate::TextureDescriptor,
-    ) -> Result<super::Texture, crate::DeviceError> {
-        let mut resource = native::Resource::null();
-
-        let raw_desc = D3D12_RESOURCE_DESC {
-            Dimension: conv::map_texture_dimension(desc.dimension),
-            Alignment: 0,
-            Width: desc.size.width as u64,
-            Height: desc.size.height,
-            DepthOrArraySize: desc.size.depth_or_array_layers as u16,
-            MipLevels: desc.mip_level_count as u16,
-            Format: if crate::FormatAspects::from(desc.format).contains(crate::FormatAspects::COLOR)
-                || !desc.usage.intersects(
-                    crate::TextureUses::RESOURCE
-                        | crate::TextureUses::STORAGE_READ
-                        | crate::TextureUses::STORAGE_READ_WRITE,
-                ) {
-                auxil::dxgi::conv::map_texture_format(desc.format)
-            } else {
-                // This branch is needed if it's a depth texture, and it's ever needed to be viewed as SRV or UAV,
-                // because then we'd create a non-depth format view of it.
-                // Note: we can skip this branch if
-                // `D3D12_FEATURE_D3D12_OPTIONS3::CastingFullyTypedFormatSupported`
-                auxil::dxgi::conv::map_texture_format_depth_typeless(desc.format)
-            },
-            SampleDesc: dxgitype::DXGI_SAMPLE_DESC {
-                Count: desc.sample_count,
-                Quality: 0,
-            },
-            Layout: d3d12::D3D12_TEXTURE_LAYOUT_UNKNOWN,
-            Flags: conv::map_texture_usage_to_resource_flags(desc.usage),
-        };
-
-        let heap_properties = d3d12::D3D12_HEAP_PROPERTIES {
-            Type: d3d12::D3D12_HEAP_TYPE_CUSTOM,
-            CPUPageProperty: d3d12::D3D12_CPU_PAGE_PROPERTY_NOT_AVAILABLE,
-            MemoryPoolPreference: match self.private_caps.memory_architecture {
-                super::MemoryArchitecture::NonUnified => d3d12::D3D12_MEMORY_POOL_L1,
-                super::MemoryArchitecture::Unified { .. } => d3d12::D3D12_MEMORY_POOL_L0,
-            },
-            CreationNodeMask: 0,
-            VisibleNodeMask: 0,
-        };
-
-        let hr = unsafe {
-            self.raw.CreateCommittedResource(
-                &heap_properties,
-                if self.private_caps.heap_create_not_zeroed {
-                    D3D12_HEAP_FLAG_CREATE_NOT_ZEROED
-                } else {
-                    d3d12::D3D12_HEAP_FLAG_NONE
-                },
-                &raw_desc,
-                d3d12::D3D12_RESOURCE_STATE_COMMON,
-                ptr::null(), // clear value
-                &d3d12::ID3D12Resource::uuidof(),
-                resource.mut_void(),
-            )
-        };
-
-        hr.into_device_result("Texture creation")?;
-        if let Some(label) = desc.label {
-            let cwstr = conv::map_label(label);
-            unsafe { resource.SetName(cwstr.as_ptr()) };
-        }
-
-        Ok(super::Texture {
-            resource,
-            format: desc.format,
-            dimension: desc.dimension,
-            size: desc.size,
-            mip_level_count: desc.mip_level_count,
-            sample_count: desc.sample_count,
-        })
-    }
-
-    #[cfg(feature = "windows_rs")]
     unsafe fn destroy_texture(&self, mut texture: super::Texture) {
         unsafe { texture.resource.destroy() };
         if let Some(alloc) = texture.allocation.take() {
-            match self.mem_allocator.lock().free(alloc) {
-                Ok(_) => (),
-                Err(e) => panic!("failed to destroy dx12 buffer, {}", e),
-            };
+            free_texture_allocation(alloc, self.mem_allocator.as_ref().unwrap());
         }
-    }
-
-    #[cfg(not(feature = "windows_rs"))]
-    unsafe fn destroy_texture(&self, texture: super::Texture) {
-        unsafe { texture.resource.destroy() };
     }
 
     unsafe fn create_texture_view(
