+++ wgpu-hal/src/dx12/mod.rs
@@ -40,6 +40,7 @@ mod descriptor;
 mod device;
 mod instance;
 mod view;
+mod windows_rs_suballocation;
 
 use crate::auxil::{self, dxgi::result::HResult as _};
 
@@ -51,6 +52,8 @@ use winapi::{
     um::{d3d12, dcomp, synchapi, winbase, winnt},
     Interface as _,
 };
+
+use self::windows_rs_suballocation::{AllocationWrapper, GpuAllocatorWrapper};
 
 #[derive(Clone)]
 pub struct Api;
@@ -238,8 +241,7 @@ pub struct Device {
     #[cfg(feature = "renderdoc")]
     render_doc: crate::auxil::renderdoc::RenderDoc,
     null_rtv_handle: descriptor::Handle,
-    #[cfg(feature = "windows_rs")]
-    mem_allocator: Mutex<gpu_allocator::d3d12::Allocator>,
+    mem_allocator: Option<Mutex<GpuAllocatorWrapper>>,
 }
 
 unsafe impl Send for Device {}
@@ -375,8 +377,7 @@ unsafe impl Sync for CommandBuffer {}
 pub struct Buffer {
     resource: native::Resource,
     size: wgt::BufferAddress,
-    #[cfg(feature = "windows_rs")]
-    allocation: Option<gpu_allocator::d3d12::Allocation>,
+    allocation: Option<AllocationWrapper>,
 }
 
 unsafe impl Send for Buffer {}
@@ -403,8 +404,7 @@ pub struct Texture {
     size: wgt::Extent3d,
     mip_level_count: u32,
     sample_count: u32,
-    #[cfg(feature = "windows_rs")]
-    allocation: Option<gpu_allocator::d3d12::Allocation>,
+    allocation: Option<AllocationWrapper>,
 }
 
 unsafe impl Send for Texture {}
@@ -771,7 +771,6 @@ impl crate::Surface<Api> for Surface {
             size: sc.size,
             mip_level_count: 1,
             sample_count: 1,
-            #[cfg(feature = "windows_rs")]
             allocation: None,
         };
         Ok(Some(crate::AcquiredSurfaceTexture {
@@ -834,39 +833,5 @@ impl crate::Queue<Api> for Queue {
         let mut frequency = 0u64;
         unsafe { self.raw.GetTimestampFrequency(&mut frequency) };
         (1_000_000_000.0 / frequency as f64) as f32
-    }
-}
-
-#[cfg(feature = "windows_rs")]
-impl From<gpu_allocator::AllocationError> for crate::DeviceError {
-    fn from(result: gpu_allocator::AllocationError) -> Self {
-        use log::error;
-
-        match result {
-            gpu_allocator::AllocationError::OutOfMemory => Self::OutOfMemory,
-            gpu_allocator::AllocationError::FailedToMap(e) => {
-                error!("DX12 gpu-allocator: Failed to map: {}", e);
-                Self::Lost
-            }
-            gpu_allocator::AllocationError::NoCompatibleMemoryTypeFound => {
-                error!("DX12 gpu-allocator: No Compatible Memory Type Found");
-                Self::Lost
-            }
-            gpu_allocator::AllocationError::InvalidAllocationCreateDesc => {
-                error!("DX12 gpu-allocator: Invalid Allocation Creation Description");
-                Self::Lost
-            }
-            gpu_allocator::AllocationError::InvalidAllocatorCreateDesc(e) => {
-                error!(
-                    "DX12 gpu-allocator: Invalid Allocator Creation Description: {}",
-                    e
-                );
-                Self::Lost
-            }
-            gpu_allocator::AllocationError::Internal(e) => {
-                error!("DX12 gpu-allocator: Internal Error: {}", e);
-                Self::Lost
-            }
-        }
     }
 }
